+++ packages/rabbitmq/src/amqp/connection.ts
@@ -109,6 +109,7 @@ const defaultConfig = {
   enableDirectReplyTo: true,
   channels: {},
   handlers: {},
+  defaultHandler: '',
   enableControllerDiscovery: false,
 };
 
@@ -424,9 +425,8 @@ export class AmqpConnection {
     consumeOptions?: ConsumeOptions
   ): Promise<SubscriptionResult> {
     return new Promise((res) => {
-      let result: SubscriptionResult;
-      this.selectManagedChannel(msgOptions?.queueOptions?.channel)
-        .addSetup(async (channel) => {
+      this.selectManagedChannel(msgOptions?.queueOptions?.channel).addSetup(
+        async (channel) => {
           const consumerTag = await this.setupSubscriberChannel<T>(
             handler,
             msgOptions,
@@ -434,11 +434,9 @@ export class AmqpConnection {
             originalHandlerName,
             consumeOptions
           );
-          result = { consumerTag };
-        })
-        .then(() => {
-          res(result);
-        });
+          res({ consumerTag });
+        }
+      );
     });
   }
 
@@ -532,20 +530,16 @@ export class AmqpConnection {
     rpcOptions: MessageHandlerOptions
   ): Promise<SubscriptionResult> {
     return new Promise((res) => {
-      let result: SubscriptionResult;
-      this.selectManagedChannel(rpcOptions?.queueOptions?.channel)
-        .addSetup(async (channel) => {
+      this.selectManagedChannel(rpcOptions?.queueOptions?.channel).addSetup(
+        async (channel) => {
           const consumerTag = await this.setupRpcChannel<T, U>(
             handler,
             rpcOptions,
             channel
           );
-          result = { consumerTag };
           res({ consumerTag });
-        })
-        .then(() => {
-          res(result);
-        });
+        }
+      );
     });
   }