+++ packages/stripe/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.2](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-stripe@0.8.1...@golevelup/nestjs-stripe@0.8.2) (2024-09-03)
+
+**Note:** Version bump only for package @golevelup/nestjs-stripe
+
 ## [0.8.1](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-stripe@0.8.0...@golevelup/nestjs-stripe@0.8.1) (2024-08-09)
 
 **Note:** Version bump only for package @golevelup/nestjs-stripe